Updated behavior to allow handling of a timeout
Prior to this update the expect method would return nil if either
the read handle were closed or a timeout occurred.  Since the
read handle being closed is clearly an error state, the code was
updated to raise an error instead of returning nil.  Now, if expect
returns nil, it should always indicate that a timeout occurred.